requirements.txt
pure-python-adb
frida